Play with ContT monad
I used to try to inherit Cont from ContT in order to reduce the noise on
the type level but it seemed to bring more problems, when I tried to
inherit the monad for Cont from the monad for ContT. Even the
composition didn't not work well: it led me to duplicate almost all the
code from the superclass. So I decided to abandon this idea and simply
start using ContT/ContTKind everywhere.

For the simple cases the internal monad is Identity. The source code
requires a lot of noise on type level, type aliases would come in handy
here.

Signed-off-by: Nikita Eshkeev <neshkeev@yandex.ru>